requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
pydantic
python-multipart
python-dotenv
tiktoken
bs4